Refactored pipeline commands to cli
