+++ docs/pipelines/status/help.go
@@ -3,5 +3,5 @@ package status
 var Usage = []string{"pl status"}
 
 func GetDescription() string {
-	return "Fetch latest pipeline run status."
+	return "Fetch the latest pipeline run status."
 }
