+++ docs/pipelines/syncstatus/help.go
@@ -1,6 +1,6 @@
 package syncstatus
 
-var Usage = []string{"pl syncstatus"}
+var Usage = []string{"pl sync status"}
 
 func GetDescription() string {
 	return "Fetch pipeline resource sync status."
