+++ docs/pipelines/version/help.go
@@ -3,5 +3,5 @@ package version
 var Usage = []string{"pl version"}
 
 func GetDescription() string {
-	return "Get pipelines version."
+	return "Show the version of JFrog Pipelines."
 }
