--- pipelines/systeminfo.go
@@ -1,18 +1,0 @@
-package pipelines
-
-import (
-	"github.com/jfrog/jfrog-cli-core/v2/common/commands"
-	status "github.com/jfrog/jfrog-cli-core/v2/pipelines/commands"
-	"github.com/urfave/cli"
-)
-
-// getVersion version command handler
-func getVersion(c *cli.Context) error {
-	serviceDetails, err := createPipelinesDetailsByFlags(c)
-	if err != nil {
-		return err
-	}
-	versionCommand := status.NewVersionCommand()
-	versionCommand.SetServerDetails(serviceDetails)
-	return commands.Exec(versionCommand)
-}
